updating daemon added
right now it (should) just auto update the daemon.app.
I'm really only deploying this for testing purposes, and
this will likely not work for those who are going to download
this version. I also added some classes to a static library
so I can share them between the cocoa application and the
auto-update application